# go/APIvidgm/docs/swagger.yaml
definitions:
  models.CDKey:
    properties:
      id:
        type: integer
      key:
        type: string
      order_item:
        $ref: '#/definitions/models.OrderItem'
      order_item_id:
        type: integer
      state:
        description: available, reserved, sold
        type: string
      user_id:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
...
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Label:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.OrderItem:
    properties:
      cdkey:
        items:
          $ref: '#/definitions/models.CDKey'
        type: array
      id:
        type: integer
      orderOwnerID:
        type: integer
      orderOwnerType:
        description: '"phy_order" o "dig_order"'
        type: string
      quantity:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
...
  models.Platform:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      modstatus:
        description: pending, approved, rejected
        type: string
      rating:
        type: integer
      user_id:
        type: integer
      username:
        type: string
      videogame_id:
        type: integer
      videogame_name:
        type: string
    type: object
...
  models.VideoGame:
    properties:
      developer:
        type: string
      dig_stock:
        type: integer
      genre:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      image_url:
        type: string
      label:
        items:
          $ref: '#/definitions/models.Label'
        type: array
      min_req:
        type: string
      pegi:
        type: integer
      phy_stock:
        type: integer
      platform:
        items:
          $ref: '#/definitions/models.Platform'
        type: array
      price:
        type: number
      rating:
        type: number
      release_year:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      sum_rating:
        type: integer
      synopsis:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Shit documentation for shit API.
  termsOfService: http://swagger.io/terms/
  title: VideoGames API
  version: "1.0"
paths:
...
  /catalog:
    get:
      consumes:
      - application/json
      description: Retrieves all entries from a specific catalog (e.g., Genre, Platform,
        or Label).
      produces:
      - application/json
      responses:
        "200":
          description: List of catalog entries
          schema:
            items:
              type: object
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all catalog entries
      tags:
      - Catalog
    post:
      consumes:
      - application/json
      description: Creates a new catalog record (e.g., Genre, Platform, or Label).
        Requires admin privileges.
      parameters:
      - description: Catalog data (Genre, Platform, or Label)
        in: body
        name: catalog
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Catalog entry created successfully
          schema:
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new catalog entry
      tags:
      - Catalog
  /catalog/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a catalog entry (e.g., Genre, Platform, or Label) by its
        ID. Requires admin privileges.
      parameters:
      - description: Catalog entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Catalog entry deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid catalog ID or entry not found
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a catalog entry
      tags:
      - Catalog
    get:
      consumes:
      - application/json
      description: Retrieves a single catalog entry (e.g., Genre, Platform, or Label)
        by its ID.
      parameters:
      - description: Catalog entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Catalog entry data
          schema:
            type: object
        "404":
          description: Catalog entry not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a catalog entry by ID
      tags:
      - Catalog
  /cdkeys:
    post:
      consumes:
      - application/json
      description: Allows an admin to create a new CD key associated with a specific
        videogame. The videogame's digital stock is incremented by 1. Requires admin
        privileges.
      parameters:
      - description: CD key data to create
        in: body
        name: cdkey
        required: true
        schema:
          $ref: '#/definitions/models.CDKey'
      produces:
      - application/json
      responses:
        "201":
          description: CD key created successfully
          schema:
            $ref: '#/definitions/models.CDKey'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden – insufficient privileges
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Videogame not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new CD key for a videogame
      tags:
      - CDKeys
  /cdkeys/{id}:
    get:
      consumes:
      - application/json
      description: Returns a CD key by ID and marks it as used if valid. Only the
        owner of the CD key can access it.
      parameters:
      - description: CD Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: CD key retrieved successfully
          schema:
            $ref: '#/definitions/models.CDKey'
        "403":
          description: Forbidden – not the owner or CD key canceled
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: CD key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Retrieve and activate a CD key
      tags:
      - CDKeys
    put:
      consumes:
      - application/json
      description: Allows an admin to update the details of an existing CD key. Requires
        admin privileges.
      parameters:
      - description: CD Key ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated CD key data
        in: body
        name: cdkey
        required: true
        schema:
          $ref: '#/definitions/models.CDKey'
      produces:
      - application/json
      responses:
        "200":
          description: CD key updated successfully
          schema:
            $ref: '#/definitions/models.CDKey'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden – insufficient privileges
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: CD key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update an existing CD key
      tags:
      - CDKeys
  /cdkeys/{id}/refund:
    patch:
      consumes:
      - application/json
      description: Allows the owner of a CD key to cancel it (refund). Only CD keys
        that are neither used nor already canceled can be refunded. Requires authentication
        as the owner.
      parameters:
      - description: CD Key ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: CD key successfully refunded (canceled)
          schema:
            $ref: '#/definitions/models.CDKey'
        "403":
          description: Forbidden – not the owner, CD key already used, or CD key without
            owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: CD key not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Refund a CD key
      tags:
      - CDKeys
...
swagger: "2.0"